+++ lib/restrr.dart
@@ -5,6 +5,7 @@ export 'src/entity_builder.dart';
 export 'src/restrr_base.dart';
 
 /* [ /src/entities ] */
+export 'src/entities/currency.dart';
 export 'src/entities/health_response.dart';
 export 'src/entities/restrr_entity.dart';
 export 'src/entities/user.dart';